folder structure update
updated the folder structure.
