+++ src/Pages/Home/actions.js
@@ -1,0 +1,6 @@
+export const getnoteList = (noteList) => {
+    return {
+        type: "GET_LIST",
+        noteList
+    }
+}
