+++ src/Pages/Login/actions.js
@@ -1,0 +1,6 @@
+export const addName = ({ name }) => {
+    return {
+        type: "ADD_USER",
+        payload: { name }
+    }
+}
