--- src/Pages/Reducers/rootReducer.js
@@ -1,9 +1,0 @@
-import notesReducer from './notes.reducer';
-import { combineReducers } from 'redux';
-
-
-const rootReducer = combineReducers({
-    notes: notesReducer,
-})
-
-export default rootReducer;
