+++ src/components/DiaryHome/actions.js
@@ -1,0 +1,6 @@
+export const addNote = ({ id, tittle, description, userName }) => {
+    return {
+        type: 'ADD_NOTE',
+        payload: { id, tittle, description, userName }
+    }
+}
